created a WerkFileUpload component
